+++ Magento/Mtf/TestCase/Injectable.php
@@ -66,6 +66,13 @@ abstract class Injectable extends Functional
     protected $currentVariation = [];
 
     /**
+     * Number of variation restart attempts.
+     *
+     * @var int
+     */
+    protected $rerunCount;
+
+    /**
      * Constructs a test case with the given name.
      *
      * @constructor
@@ -79,6 +86,7 @@ abstract class Injectable extends Functional
         parent::__construct($name, $data, $dataName);
         $this->dataId = get_class($this) . '::' . $name;
         $this->filePath = $path;
+        $this->rerunCount = empty($_ENV['rerun_count']) ? 0 : $_ENV['rerun_count'];
     }
 
     /**
@@ -156,7 +164,12 @@ abstract class Injectable extends Functional
                     $this->localArguments
                 );
                 $this->executeTestVariation($result, $variation);
-                $testVariationIterator->next();
+                if ($this->rerunCount > 0 && $this->getStatus() != \PHPUnit_Runner_BaseTestRunner::STATUS_PASSED) {
+                    $this->rerunCount -= 1;
+                } else {
+                    $this->rerunCount = empty($_ENV['rerun_count']) ? 0 : $_ENV['rerun_count'];
+                    $testVariationIterator->next();
+                }
                 $this->localArguments = [];
             }
         } catch (\PHPUnit_Framework_IncompleteTestError $phpUnitException) {
@@ -207,6 +220,7 @@ abstract class Injectable extends Functional
         if (isset($this->currentVariation['arguments']['issue'])
             && !empty($this->currentVariation['arguments']['issue'])
         ) {
+            $this->rerunCount = 0;
             $this->markTestIncomplete($this->currentVariation['arguments']['issue']);
         }
         $testResult = parent::runTest();
@@ -261,9 +275,9 @@ abstract class Injectable extends Functional
             $arguments = array_merge($variation['arguments'], $arguments);
         }
         if (isset($variation['arguments']['variation_name'])) {
-            $this->setVariationName($variation['arguments']['variation_name']);
+            $this->setVariationName($variation['arguments']['variation_name'] . "_" . $this->rerunCount);
         } else {
-            $this->setVariationName($variation['id']);
+            $this->setVariationName($variation['id'] . "_" . $this->rerunCount);
         }
         $resolvedArguments = $this->getObjectManager()
             ->prepareArguments($this, $this->getName(false), $arguments);